Refactored for readability + WIP Home
