+++ .prettierrc
@@ -2,5 +2,6 @@
     "trailingComma": "es5",
     "tabWidth": 4,
     "semi": true,
+    "plugins": ["prettier-plugin-css-order"],
     "singleQuote": true
 }
