+++ package.json
@@ -91,7 +91,8 @@
     },
     "devDependencies": {
         "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-        "prettier": "3.0.1"
+        "prettier": "3.0.1",
+        "prettier-plugin-css-order": "^2.0.0"
     },
     "jest": {
         "roots": [
