+++ src/Components/dropdown.module.css
@@ -3,22 +3,22 @@
 }
 
 .dropdown__dropdownMenu {
+    display: flex;
+    position: absolute;
+    top: calc(100% + 3px);
+    right: 0;
+    row-gap: 0.5rem;
+    flex-direction: column;
     visibility: hidden;
-    will-change: transform;
     opacity: 0;
-    min-width: 160px;
-    text-align: center;
+    will-change: transform;
     box-shadow: rgba(0, 0, 0, 0.3) 0 2px 2px 0;
     border-radius: 3px;
-    position: absolute;
-    right: 0;
-    display: flex;
-    flex-direction: column;
-    row-gap: 0.5rem;
-    padding: 0.4rem;
-    top: calc(100% + 3px);
     background: var(--primary-text-color);
+    padding: 0.4rem;
+    min-width: 160px;
     color: var(--text-neutral-color);
+    text-align: center;
 }
 
 .dropdown__dropdownMenu_visible {
@@ -27,17 +27,17 @@
 }
 
 .dropdown__dropdownMenuItem {
-    background: var(--primary-text-color);
-    cursor: pointer;
-    border: none;
-    border-radius: 3px;
-    padding: 0.5rem 1rem;
     display: flex;
     justify-content: space-between;
     align-items: center;
+    cursor: pointer;
+    border: none;
+    border-radius: 3px;
+    background: var(--primary-text-color);
+    padding: 0.5rem 1rem;
     min-width: 100%;
-    font-size: 0.8rem;
     color: var(--primary-color);
+    font-size: 0.8rem;
 }
 
 .dropdown__dropdownMenuItem:focus {
