+++ src/Layout/Base.tsx
@@ -1,11 +1,14 @@
 import { Outlet } from 'react-router-dom';
-import Navbar from 'src/Components/Shared/Navbar';
+import Navbar from 'src/Layout/Shared/Navbar';
+import styles from './base.module.css';
 
 const Base = () => {
     return (
         <>
             <Navbar />
-            <Outlet />
+            <div className={styles.baseLayout__content}>
+                <Outlet />
+            </div>
         </>
     );
 };
