+++ src/Layout/base.module.css
@@ -1,0 +1,4 @@
+.baseLayout__content {
+    margin: 0 auto;
+    max-width: var(--max-content-width);
+}
